Add initial engineer recycling
With recycling we refer to evaluating the state machine of each engineer. If the state machine is considered to be in a final state ('Completed' or 'Error') then we re-assign the behavior of the engineer to the idle behavior.

This is a starting point of a loop to manage engineers!